% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{common_headers}
\alias{common_headers}
\title{common_headers}
\format{
An object of class \code{character} of length 10.
}
\usage{
common_headers
}
\description{
common_headers
}
\keyword{datasets}
