% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ctg_cookies}
\alias{ctg_cookies}
\title{ctg_cookies}
\format{
An object of class \code{character} of length 1.
}
\usage{
ctg_cookies
}
\description{
A template for cleaningtheglass.com cookies.
This is a placeholder - you will need to provide your own session ID
by replacing 'YOUR_SESSION_ID' with your actual session ID from the website
}
\keyword{datasets}
