% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-package.R
\name{fill_desc}
\alias{fill_desc}
\title{Initialize DESCRIPTION.}
\usage{
fill_desc(pkgname, email, title, description)
}
\arguments{
\item{pkgname}{Name of the package}

\item{email}{Email address for the package author}

\item{title}{Package title}

\item{description}{Description of the R package.}
}
\value{
DESCRIPTION file populated with default and user-supplied values.
}
\description{
Initialize DESCRIPTION.
}
\examples{
\dontrun{
fill_desc("mypackage", "example@example.com", "An example R package",
"This is an example R package.")
}
}
