% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_season_strings}
\alias{generate_season_strings}
\title{generate_season_strings}
\usage{
generate_season_strings(start_year, end_year)
}
\arguments{
\item{start_year}{start_year}

\item{end_year}{end_year}
}
\value{
seasons
}
\description{
generate_season_strings
}
