% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_formula_from_model}
\alias{get_formula_from_model}
\title{get_formula_from_model}
\usage{
get_formula_from_model(model, df_name)
}
\arguments{
\item{model}{model}

\item{df_name}{df_name}
}
\value{
formula
}
\description{
get_formula_from_model
}
