% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_html_table_from_url}
\alias{get_html_table_from_url}
\title{get_html_table_from_url}
\usage{
get_html_table_from_url(url, idx = 1)
}
\arguments{
\item{url}{url}

\item{idx}{idx}
}
\value{
df
}
\description{
get_html_table_from_url
}
