% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_nba_com_table}
\alias{get_nba_com_table}
\title{get_nba_com_table}
\usage{
get_nba_com_table(url, params, idx = 1, ignore_ranks = F)
}
\arguments{
\item{url}{url}

\item{params}{params}

\item{idx}{idx}

\item{ignore_ranks}{ignore_ranks}
}
\value{
df
}
\description{
get_nba_com_table
}
