% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pbp_stats_table}
\alias{get_pbp_stats_table}
\title{get_pbp_stats_table}
\usage{
get_pbp_stats_table(url, headers, params, single_row = F)
}
\arguments{
\item{url}{url}

\item{headers}{headers}

\item{params}{params}

\item{single_row}{single_row}
}
\value{
df
}
\description{
get_pbp_stats_table
}
