% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_team_color}
\alias{get_team_color}
\title{get_team_color}
\usage{
get_team_color(
  team,
  color = df_teams$COLOR_PRIMARY[df_teams$ABV \%in\% team],
  secondary = F
)
}
\arguments{
\item{team}{team}

\item{color}{color}

\item{secondary}{secondary}
}
\value{
color
}
\description{
get_team_color
}
