% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-package.R
\name{init_data_raw}
\alias{init_data_raw}
\title{Initialize a raw data folder and R scripts for setting up a package and munging the raw data.}
\usage{
init_data_raw(name = "pkg-devops", data_title = "01-load")
}
\arguments{
\item{name}{String naming an R script for package initialization}

\item{data_title}{String naming an R script for loading raw data}
}
\value{
A \code{data-raw} folder and two files with names based on the arguments supplied to the function. The data loading script will be opened for editing.
}
\description{
Initialize a raw data folder and R scripts for setting up a package and munging the raw data.
}
\examples{
\dontrun{
init_data_raw()
}
}
