% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-package.R
\name{init_docs}
\alias{init_docs}
\title{Initialize documentation files for an R package, including tests, license, news, and roxygen markdown.}
\usage{
init_docs(auth_name = "Frank Farach", testthat = TRUE,
  license = TRUE, news = TRUE, roxygen_md = TRUE)
}
\arguments{
\item{auth_name}{Character string naming the package author}

\item{testthat}{Logical flag indicating whether to include testing}

\item{license}{Logical flag indicating whether to include a \code{LICENSE} file}

\item{news}{Logical flag indicating whether to include \code{NEWS.md}}

\item{roxygen_md}{Logical flag indicating whether to use the \code{roxygen_md} package}
}
\value{
The return values depend on the flags that are TRUE.
}
\description{
Initialize documentation files for an R package, including tests, license, news, and roxygen markdown.
}
\examples{
\dontrun{
init_docs()
}
}
