% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_counter}
\alias{make_counter}
\title{Create a counter function.}
\usage{
make_counter(init = 0, count_by = 1)
}
\arguments{
\item{init}{An initial number for the counter.}

\item{count_by}{A number for skip counting.}
}
\value{
A counter function with a counter initialized to \code{init} that increments by \code{count_by} each time the function is called.
}
\description{
Create a counter function.
}
\examples{
counter <- make_counter()
counter()
counter()

count_from_1 <- make_counter(init = 1)
count_from_1()
count_from_1()

count_down_from_10 <- make_counter(count_by = -1)
count_down_from_10()
count_down_from_10()
}
