% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{nba_com_season_types}
\alias{nba_com_season_types}
\title{nba_com_season_types}
\format{
An object of class \code{character} of length 5.
}
\usage{
nba_com_season_types
}
\description{
nba_com_season_types
}
\keyword{datasets}
