% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{pbp_stats_headers}
\alias{pbp_stats_headers}
\title{pbp_stats_headers}
\format{
An object of class \code{character} of length 12.
}
\usage{
pbp_stats_headers
}
\description{
pbp_stats_headers
}
\keyword{datasets}
