% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preview}
\alias{preview}
\title{Preview function}
\usage{
preview(df, n = 3)
}
\arguments{
\item{df}{A data frame to preview}

\item{n}{Number of rows to show from beginning and end (default is 3)}
}
\description{
Quickly view the first and last n rows of a data frame in RStudio's viewer.
This is useful for getting a quick glimpse of large datasets.
}
\details{
This function opens two views in RStudio's Viewer: one showing the first n rows
and another showing the last n rows of the provided data frame. The function captures
the variable name dynamically using system calls.
}
\examples{
# Using with mtcars dataset
preview(mtcars)  # Shows first 3 and last 3 rows

# Using with 5 rows
# preview(mtcars, n = 5)

}
