% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remaining_sample_avg}
\alias{remaining_sample_avg}
\title{remaining_sample_avg}
\usage{
remaining_sample_avg(entire_weight, entire_value, sample_weight, sample_value)
}
\arguments{
\item{entire_weight}{entire_weight}

\item{entire_value}{entire_value}

\item{sample_weight}{sample_weight}

\item{sample_value}{sample_value}
}
\value{
avg
}
\description{
remaining_sample_avg
}
