% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{square_logo_base_url}
\alias{square_logo_base_url}
\title{square_logo_base_url}
\format{
An object of class \code{character} of length 1.
}
\usage{
square_logo_base_url
}
\description{
square_logo_base_url
}
\keyword{datasets}
