% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_nba_com_table}
\alias{get_nba_com_table}
\title{get_nba_com_table}
\usage{
get_nba_com_table(url, headers, params, idx = 1)
}
\arguments{
\item{url}{url}

\item{headers}{headers}

\item{params}{params}

\item{idx}{idx}
}
\value{
df
}
\description{
get_nba_com_table
}
